requirements.txt
asgiref==3.3.1
Babel==2.9.0
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
django-admin==2.0.1
django-appconf==1.0.4
django-crispy-forms==1.11.0
django-debug-toolbar==3.2
django-environ==0.4.5
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-select2==7.6.1
django-settings-export==1.2.1
django-six==1.0.4
djangorestframework==3.12.2
i18n==0.2
idna==2.10
mysqlclient==2.0.3
phonenumbers==8.12.18
Pillow==8.1.0
pkg-resources==0.0.0
py==1.10.0
python-gettext==4.0
pytz==2021.1
requests==2.25.1
screen==1.0.1
sqlparse==0.4.1
urllib3==1.26.3
xlwt==1.3.0
